Copy over game files
This is a large commit - the files were previously part of a private repo.

This game is a WIP, so far I have the initial set of cards I want to implement and a basic rendering engine for rendering cards.  I've identified some of the technologies I'm going to use to build it (such as socket.io)

Next up are:
- copying over my CI files so that we can run github CI on each new PR / commit that gets pushed

- writing a readme file to explain everything