Add room joining demo
Part of #9

Adding a basic demo for a 'room' that a user can join after selecting a name.

When users join these rooms (which are powered by socket.io's rooms feature), they update all the other sockets/users that someone joined a room.

Similarly, when a user leaves a room or disconnects, or resets their name, it disconnects them from the room and makes them update everyone else.

I chose this approach b/c I'm thinking the number of users may be small and we can tackle the scaling issues of the game later.  I just wanted to get something working and functional for game lobbies.

Added a bunch of todo's like using adapters for game logic, adding limits to number of rooms that can be joined, etc.  We'll tackle this + the idea of 'starting a game' and 'making game actions' in upcoming PR's