Make type.d.ts a non-ambient file
I'm adding a serverToClientEvent method called updateBoard that updates sockets with a particular board state.

In order to integrate this, we need to add an import file to types.d.ts.

However, the moment that we have an import in a .d.ts file, it no longer becomes global:

https://stackoverflow.com/questions/39040108/import-class-in-definition-file-d-ts

Renamed it to just types.ts to reflect this