Reduce time to first paint by 25% (SEO improvements)
I recently added CloudFlare as a CDN provider for the app and tackling the Google Lighthouse issues highlighted in #157 .

The biggest problem highlighted by the report was the client-side Javascript resources, resulting in a 4.8s First Contentful Paint + time to interactive.  On 3G this becomes even worse, with a 10.5second first contentful paint.  (Source: lighthouse report, see #158 )

Added code splitting + bundle minimization to reduce the overall size of client.bundle.js.

Every piece of the app besides the router was put into bundles that load when actually called.

This reduced our overall bundle size from 274kb to 217kb
Before:

After:

The impact on speed was also signifcant

Before (3.0 seconds time to first paint):

After: (2.4 seconds time to first paint):

B/c file splitting / lazy-load causes the app to paint in a jarring manner, I introduced an animation for the appearance of the NameChanger component to reduce strain on the eyes:

Borrowed inspiration from:
https://nextjs.org/docs/advanced-features/dynamic-import#with-named-exports