Add keyboard a11y
Big accessibility for keyboard-only users.  The whole game can be played sans mouse now.

Added tab-indexes

TabIndex's are a good thing (this is a11y 101) => your users should be able to navigate your entire site sans-mouse for a11y reasons and also personal ergonomics
A problem arises when you want to do tabIndexes and let users use the Enter key to press on divs (yes, I know buttons/links are better, but sometimes we really do need clickable divs).
To that end I followed this stack overflow answer.  I improved upon it by using keyup instead of keypress which is getting deprecated:
        <script>
            const handleEnter = (event) => {
                if (
                    event.key === 'Enter' &&
                    document.activeElement instanceof HTMLElement&&
                    document.activeElement.tagName === 'div' // need to do this to prevent double-clicks on buttons + links
                ) {
                    document.activeElement.click();
                }
            };
        </script>

    </head>
    <body onkeyup="handleEnter(event)">
This is much easier than having to memoize the handleEnter at the top of the React app and then passing it down there to the first div in the node tree.
4. You have to ensure that the place that you have your onClick on the same element that you have your tabIndex and not a parent element with this hack (as onClick's bubble up, but not down)
5. over in Vue land, it seems like they have some sort of mix-ins that they can use on their elements like this: <div tabindex="0" @keydown.enter="handleClick" @click="handleClick"> (source) (edited)