Feat: Bump decklist size to 60
Originally, I had wanted the game to feel "unique" from Magic (which has a 60 card deck) and I had landed on a 48-card deck size.

The problem with 48 cards is that it makes decks feel repetitive when you can have 4-of's in a deck.  It also doesn't allow the whole card-pool to shine as much.

- Bumped the deck minimum size to 60
- Made the dual lands that come in untapped deal 2 damage to the user
- Added a "pirates" deck