Fix issue where effects would linger without a valid target
Found this while play-testing.

Closes: #360

We had an issue where effects (like Old Farmhouse's "buff attack" effect) would linger on the game client if they didn't have a target type specified (we use `getDefaultTargetForEffect` for these types of effects).

Fixed this by adding the logic in for grabbing the default targets in the useEffect that handles fizzling the effects that don't have valid targets.