Prevent tooltip from overflowing
New discovery: If tooltips are too long to fit on the top of the page, popper will try to place them at the bottom, causing overflow / scroll issues.

Popper is essentiallly giving up and saying, hey there's no good way to place this item, so let's place it at the bottom.

Fixed by shortening the help text on the popup