Helpers for creating a "drafting" / "sealed" experience
In this game, everything in the game that is displayed derives from the state of the Board object, as passed from the server, calling the obscureBoardInfo helper to obscure critical pieces of information (e.g. opponent's hand, etc.).

Game actions from the client are sent as actions to the server, which calculate a potential new game board state to display/

When approaching the challenge of building "draft" mode, a logical place to start is to modify and expand the concept of the "Board" objects to encapsulate everything to do with drafting - draft pool, draft piles, etc.

Players also get an update in that they have a `deckBuildingPool` property as well, which starts as a blank array (or an array of 6 packs for sealed mode).

This PR upgraded the makeNewBoard / makeNewPlayer to account for sealed and draft mode games.