Feat: Future proof game for sideboards
As I started working on draft mode, I thought it would be nice to support sideboards in the future in case players wanted to play best of 3 or even best of 5.

This update doesn't add sideboards in per se, but it does allow for the possibility of it in the future.

With Typescript, it was very easy to change the structure of `DeckList` and `Skeleton` to take a new structure.

The rest was paint-by-the-numbers using jest:watch to fix all the places where everything needed updating.