Feat: Nerf / buff cards
After playtesting - a few cards need nerfing.

Curious researcher feels like a better version of a 2/2 draw a card for 3.  It has too much upside for no downside, so nerfing it to 4 mana.

Lu Zishen is too strong as a 3/3 - since it hits twice, it can deal a whopping 6 damage when ahead, while also taking out 4 damage worth of opposing units

Penelope the Tortoise receives a buff - healing for 4, drawing one less card (3 -> 2)

The 'draw per unit' cards, Merry Rallier and Team Gathering, have been made to cost 1 less.  These cards are difficult currently to get to in terms of mana cost compared to their upside, and need to be slightly more accessible