Feat: Add "passive" effect requirements
Adds effect requirements that are passive in nature e.g. 'control a legendary leader' or 'have at least 3 cards in cemetery'.  This is opposed to an 'active requirement', e.g. "do this only if you discard a resource card".  The distinction between the two types of effect requirements for now is purely for organizing the code.

Also: adds a feature to buff the team's generic units.

This is the basis for several new upcoming cards